code clean up

+++ data/movies/src/main/java/com/example/movies/paging/MoviesRemoteMediator.kt
@@ -1,6 +1,5 @@
 package com.example.movies.paging
 
-import android.util.Log
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
@@ -8,73 +7,61 @@ import androidx.paging.rxjava2.RxRemoteMediator
 import com.example.database.ILocalDataSource
 import com.example.database.entities.MovieEntity
 import com.example.database.entities.MovieEntityRemoteKey
-import com.example.network.movies.IMoviesNetworkDataSource
+import com.example.network.INetworkDataSource
 import com.example.util.toEntity
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
 
+private const val LAST_PAGE = 78
+private const val DEFAULT_PAGE = 1
+
 @OptIn(ExperimentalPagingApi::class)
 class MoviesRemoteMediator(private val localSource: ILocalDataSource,
-                           private val remoteSource: IMoviesNetworkDataSource): RxRemoteMediator<Int, MovieEntity>() {
+                           private val remoteSource: INetworkDataSource
+): RxRemoteMediator<Int, MovieEntity>() {
 
     override fun loadSingle(
         loadType: LoadType,
         state: PagingState<Int, MovieEntity>
     ): Single<MediatorResult> {
 
-        var remoteKeySingle: Single<MovieEntityRemoteKey>? = null
-        remoteKeySingle = when (loadType) {
+        val remoteKeySingle: Single<MovieEntityRemoteKey> = when (loadType) {
 
-            LoadType.REFRESH -> {
-                Log.e("JEFF", "REFRESH")
-                Single.just(MovieEntityRemoteKey(1, 1))
-            }
+            LoadType.REFRESH -> Single.just(MovieEntityRemoteKey(DEFAULT_PAGE, DEFAULT_PAGE))
 
             LoadType.PREPEND -> return Single.just(MediatorResult.Success(true))
 
-            LoadType.APPEND -> {
-                Log.e("JEFF", "APPEND")
-                getRemoteKeyForLastMovieItemOnTheList(state)
-            }
+            LoadType.APPEND -> getRemoteKeyForLastMovieItemOnTheList(state)
         }
 
-        val x = remoteKeySingle.subscribeOn(Schedulers.io()).flatMap { remoteKey ->
+        return remoteKeySingle.subscribeOn(Schedulers.io()).flatMap { remoteKey ->
 
-                if (loadType != LoadType.REFRESH && remoteKey.nextKey == 78) {
-                    Log.e("JEFF", "111111111111")
+                if (loadType != LoadType.REFRESH && remoteKey.nextKey == LAST_PAGE) {
                     return@flatMap Single.just<MediatorResult>(MediatorResult.Success(true))
                 }
 
                 remoteSource.getMovies(remoteKey.nextKey).map { response ->
 
                     localSource.database().runInTransaction {
-                        if (loadType == LoadType.REFRESH) {
-                            localSource.deleteAll()
-                        }
+                        if (loadType == LoadType.REFRESH) { localSource.deleteAll() }
 
-                        Log.e("TOSIN", "HERE ${response.page}")
-                        localSource.insertOrReplace(MovieEntityRemoteKey(response.page, response.page + 1))
-                        localSource.insertAll(response.results.toEntity(response.page))
+                        localSource.insert(MovieEntityRemoteKey(response.page, response.page + 1))
+                        localSource.insert(response.results.toEntity(response.page))
                     }
 
-                    Log.e("JEFF", "2222222 ${response.page == 78}")
-                    return@map MediatorResult.Success(response.page == 78)
+                    return@map MediatorResult.Success(response.page == LAST_PAGE)
                 }
 
-            }.onErrorResumeNext {e ->
-            Log.e("JEFF", "3333333")
+            }.onErrorResumeNext { e ->
                 return@onErrorResumeNext Single.just(MediatorResult.Error(e))
             }
-
-        return x
     }
 
     private fun getRemoteKeyForLastMovieItemOnTheList(state: PagingState<Int, MovieEntity>): Single<MovieEntityRemoteKey> {
         val lastMovieEntityOnTheList = state.lastItemOrNull()
-            ?: return Single.just(MovieEntityRemoteKey(1, 1))
+            ?: return Single.just(MovieEntityRemoteKey(DEFAULT_PAGE, DEFAULT_PAGE))
 
-        val remoteKeyForLastMovieEntityOnTheList = localSource.remoteKeyByQuery(lastMovieEntityOnTheList.pageNumber)
-        return remoteKeyForLastMovieEntityOnTheList
+        return localSource.getRemoteKey(lastMovieEntityOnTheList.pageNumber)
     }
 
 }
